flask-server/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.0.1
gunicorn==20.1.0
matplotlib==3.3.4
tensorflow_addons==0.13.0
opencv_python==4.5.1.48
keras_nightly==2.5.0.dev2021032900
numpy==1.19.5
tensorflow-cpu==2.5.0rc0
torch==1.7.1+cpu
torchvision==0.8.2+cpu
pandas==1.2.3
keras==2.4.3
Pillow==8.3.1
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.41.0
tensorboard>=2.4.1
seaborn>=0.11.0
thop